#4 libarray stat median function

+++ kochan/libcommon/libari.h
@@ -208,9 +208,9 @@ bool ari_range_isequal(int a[], int na, int rstarta, int renda,
 // calculate average
 double ari_stat_mean(int a[], int n);
 // calculate median
-int    ari_stat_median(int a[], int n, int median[2]);
+double ari_stat_median(int a[], int n, int median[2]);
 // calculate modes
-int    ari_stat_mode(int a[], int n, int mode[], int freq[]);
+int ari_stat_mode(int a[], int n, int mode[], int freq[]);
 // standard deviation
 double ari_stat_stddev(int a[], int n);
 // variance
